Make eligibility based on workspace owner's initial credit eligibility.
