+++ api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsService.java
@@ -1,6 +1,5 @@
 package org.pmiops.workbench.initialcredits;
 
-import static java.time.temporal.ChronoUnit.DAYS;
 import static org.pmiops.workbench.db.dao.WorkspaceDao.WorkspaceCostView;
 
 import jakarta.annotation.Nullable;
@@ -10,7 +9,6 @@ import jakarta.validation.constraints.NotNull;
 import java.sql.Timestamp;
 import java.time.Clock;
 import java.time.Duration;
-import java.time.Instant;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.Comparator;
@@ -21,7 +19,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
-import org.mapstruct.Named;
 import org.pmiops.workbench.actionaudit.auditors.UserServiceAuditor;
 import org.pmiops.workbench.cloudtasks.TaskQueueService;
 import org.pmiops.workbench.config.WorkbenchConfig;
@@ -391,28 +388,6 @@ public class InitialCreditsService {
                     workbenchConfigProvider.get().billing.initialCreditsExtensionPeriodDays)));
     userInitialCreditsExpiration.setExtensionTime(clockNow());
     return userDao.save(user);
-  }
-
-  @Named("checkInitialCreditsExtensionEligibility")
-  public boolean checkInitialCreditsExtensionEligibility(DbUser dbUser) {
-    DbUserInitialCreditsExpiration initialCreditsExpiration =
-        dbUser.getUserInitialCreditsExpiration();
-    Instant now = Instant.now();
-    WorkbenchConfig.BillingConfig billingConfig = workbenchConfigProvider.get().billing;
-
-    return initialCreditsExpiration != null
-        && initialCreditsExpiration.getExtensionTime() == null
-        && initialCreditsExpiration.getCreditStartTime() != null
-        && now.isAfter(
-            initialCreditsExpiration
-                .getExpirationTime()
-                .toInstant()
-                .minus(billingConfig.initialCreditsExpirationWarningDays, DAYS))
-        && now.isBefore(
-            initialCreditsExpiration
-                .getCreditStartTime()
-                .toInstant()
-                .plus(billingConfig.initialCreditsExtensionPeriodDays, DAYS));
   }
 
   private void checkExpiration(DbUser user) {
